feat(settings): Add ISO certification link to company settings
Introduces a new field `isoLinkUrl` to the `CompanyInfo` type. This allows users to specify a URL for their ISO certification, which will be displayed as a clickable link alongside the ISO logo in the company footer. The admin interface has been updated to include an input field for this URL.